+++ changelog.md
@@ -3,6 +3,7 @@
   - Mods tab: right-click on a version number for a menu to recheck updates.
   - Mods tab: option to need a double-click to open the side panel.
     - Default is to need double-click (currently is single-click).
+  - Debug feature to update to a specific release.
 - Fixed
   - Changelogs not shown correctly for mods using Google Drive to host the text. 
   - Double scrollbar on Mods tab.